+++ wgpu-core/src/command/transfer.rs
@@ -666,7 +666,10 @@ fn handle_texture_init(
         kind: init_kind,
     };
 
-    // Register the init action.
+    // Record the initialization action. Simultaneously, collect a list of any
+    // ranges of the texture that were discarded within the current command
+    // buffer, for immediate initialization. (The analogous case for passes is
+    // in `fixup_discarded_surfaces`.)
     let immediate_inits = state
         .texture_memory_actions
         .register_init_action(&{ init_action });
@@ -674,12 +677,20 @@ fn handle_texture_init(
     // In rare cases we may need to insert an init operation immediately onto the command buffer.
     if !immediate_inits.is_empty() {
         for init in immediate_inits {
+            let index = init.layer_or_depth_slice;
+            let (layer_range, depth_slice) = if texture.desc.dimension == wgt::TextureDimension::D3
+            {
+                (0..1, Some(index))
+            } else {
+                (index..(index + 1), None)
+            };
             clear_texture(
                 &init.texture,
                 TextureInitRange {
                     mip_range: init.mip_level..(init.mip_level + 1),
-                    layer_range: init.layer..(init.layer + 1),
+                    layer_range,
                 },
+                depth_slice,
                 state.raw_encoder,
                 &mut state.tracker.textures,
                 &state.device.alignments,